use base_llm_http_handler for cohere
